fix: remove application and user roles components
